% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QKResults_class.R
\name{new_QKResults}
\alias{new_QKResults}
\title{QKResult Constructor}
\usage{
new_QKResults(qkList)
}
\arguments{
\item{qkList}{list(quants, yquants, g, l, ll <- -optparb$value, beta0, nu, xstar, ystar, Ki, quantv, mult, ylisto, type)}
}
\value{
New class QKResults
}
\description{
Create Quantile Kriging Results class from list
}
